+++ controlePedidos/src/main/java/com/br/chocolatePontoVirgula/model/repository/ProdutoRepository.java
@@ -4,7 +4,7 @@ import com.br.chocolatePontoVirgula.model.entity.Produto;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 
 
 }
